Bug: error on console
There is another warning in console, The current version of "react-router-dom" creates the warning in react.
A future version may fix it but for now if we want to fix it there is a need to change it ourselves(copying the react-router-dom library and try to ignore the warning in restricted mode) which is not possible by me.
All our links use this library, I suggest we wait for a stable update for this library the current more updated library is unstable so its not wroth the risk of updating.